+++ src/main/kotlin/app/revanced/patches/tiktok/misc/settings/patch/TikTokSettingsPatch.kt
@@ -6,24 +6,20 @@ import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
+import app.revanced.patcher.extensions.InstructionExtensions.addInstructionsWithLabels
 import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
-import app.revanced.patcher.extensions.InstructionExtensions.replaceInstruction
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchResult
-import app.revanced.patcher.patch.PatchResultError
 import app.revanced.patcher.patch.PatchResultSuccess
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patcher.patch.annotations.Patch
+import app.revanced.patcher.util.smali.ExternalLabel
 import app.revanced.patches.tiktok.misc.integrations.patch.TikTokIntegrationsPatch
 import app.revanced.patches.tiktok.misc.settings.annotations.TikTokSettingsCompatibility
-import app.revanced.patches.tiktok.misc.settings.fingerprints.AboutPageFingerprint
-import app.revanced.patches.tiktok.misc.settings.fingerprints.AdPersonalizationActivityOnCreateFingerprint
-import app.revanced.patches.tiktok.misc.settings.fingerprints.SettingsOnViewCreatedFingerprint
+import app.revanced.patches.tiktok.misc.settings.fingerprints.*
 import org.jf.dexlib2.Opcode
-import org.jf.dexlib2.iface.instruction.FiveRegisterInstruction
-import org.jf.dexlib2.iface.instruction.OneRegisterInstruction
-import org.jf.dexlib2.iface.instruction.ReferenceInstruction
-import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
+import org.jf.dexlib2.builder.instruction.BuilderInstruction35c
+import org.jf.dexlib2.iface.instruction.formats.Instruction35c
 
 @Patch
 @DependsOn([TikTokIntegrationsPatch::class])
@@ -33,65 +29,46 @@ import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
 @Version("0.0.1")
 class TikTokSettingsPatch : BytecodePatch(
     listOf(
-        AboutPageFingerprint,
         AdPersonalizationActivityOnCreateFingerprint,
-        SettingsOnViewCreatedFingerprint,
+        AddSettingsEntryFingerprint,
+        SettingsEntryFingerprint,
+        SettingsEntryInfoFingerprint,
     )
 ) {
     override fun execute(context: BytecodeContext): PatchResult {
-        SettingsOnViewCreatedFingerprint.result?.mutableMethod?.apply {
-            val instructions = implementation!!.instructions
+        // Find the class name of classes which construct a settings entry
+        val settingsButtonClass = SettingsEntryFingerprint.result?.classDef?.type?.toClassName()
+            ?: return SettingsEntryFingerprint.toErrorResult()
+        val settingsButtonInfoClass = SettingsEntryInfoFingerprint.result?.classDef?.type?.toClassName()
+            ?: return SettingsEntryInfoFingerprint.toErrorResult()
 
-            // Find the indices that need to be patched.
-            val copyrightPolicyLabelId = AboutPageFingerprint.result?.let {
-                val startIndex = it.scanResult.patternScanResult!!.startIndex
-                it.mutableMethod.getInstruction<WideLiteralInstruction>(startIndex).wideLiteral
-            } ?: return AboutPageFingerprint.toErrorResult()
-
-            val copyrightIndex = instructions.indexOfFirst {
-                (it as? ReferenceInstruction)?.reference.toString() == "copyright_policy"
-            } - 6
-
-
-            // fixme: instead use Method.indexOfFirstConstantInstructionValue()
-            val copyrightPolicyIndex = instructions.indexOfFirst {
-                (it as? WideLiteralInstruction)?.wideLiteral == copyrightPolicyLabelId
-            } + 2
-
-            // Replace an existing settings entry with ReVanced settings entry.
-            arrayOf(
-                copyrightIndex,
-                copyrightPolicyIndex
-            ).forEach { index ->
-                val instruction = getInstruction(index)
-                if (instruction.opcode != Opcode.MOVE_RESULT_OBJECT)
-                    return PatchResultError("Hardcoded offset changed.")
-
-                val settingsEntryStringRegister = (instruction as OneRegisterInstruction).registerA
-
-                // Replace the settings entry string with a custom one.
-                replaceInstruction(
-                    index,
-                    """
-                    const-string v$settingsEntryStringRegister, "ReVanced Settings"
-                """
+        // Create a settings entry for 'revanced settings' and add it to settings fragment
+        AddSettingsEntryFingerprint.result?.apply {
+            scanResult.patternScanResult?.startIndex?.let {
+                val settingsEntries = mutableMethod.getInstruction(it + 3)
+                val addEntry = mutableMethod.getInstruction<BuilderInstruction35c>(it + 5)
+                val register1 = addEntry.registerC
+                val register2 = addEntry.registerD
+                // Add the settings entry created to the settings fragment
+                mutableMethod.addInstructions(
+                    it + 6,
+                    listOf(
+                        settingsEntries,
+                        addEntry
+                    )
                 )
-
-                // Replace the OnClickListener class with a custom one.
-                val onClickListener = getInstruction<ReferenceInstruction>(index + 4).reference.toString()
-
-                context.findClass(onClickListener)?.mutableClass?.methods?.first {
-                    it.name == "onClick"
-                }?.addInstructions(
-                    0,
+                // These instructions call a method that create a settings entry use reflection base on the class name of classes that construct settings entry
+                mutableMethod.addInstructions(
+                    it + 6,
                     """
-                        invoke-static {}, $INTEGRATIONS_CLASS_DESCRIPTOR->startSettingsActivity()V
-                        return-void
+                        const-string v$register1, "$settingsButtonClass"
+                        const-string v$register2, "$settingsButtonInfoClass"
+                        invoke-static {v$register1, v$register2}, $CREATE_SETTINGS_ENTRY_METHOD_DESCRIPTOR
+                        move-result-object v$register2
                     """
-                ) ?: return PatchResultError("Could not find the onClick method.")
+                )
             }
-
-        } ?: return SettingsOnViewCreatedFingerprint.toErrorResult()
+        } ?: return AddSettingsEntryFingerprint.toErrorResult()
 
         // Initialize the settings menu once the replaced setting entry is clicked.
         AdPersonalizationActivityOnCreateFingerprint.result?.mutableMethod?.apply {
@@ -99,18 +76,26 @@ class TikTokSettingsPatch : BytecodePatch(
                 it.opcode == Opcode.INVOKE_SUPER
             } + 1
 
-            val thisRegister = getInstruction<FiveRegisterInstruction>(initializeSettingsIndex - 1).registerC
+            val thisRegister = getInstruction<Instruction35c>(initializeSettingsIndex - 1).registerC
+            val usableRegister = implementation!!.registerCount - parameters.size - 2
 
-            addInstructions(
+            addInstructionsWithLabels(
                 initializeSettingsIndex,
                 """
                     invoke-static {v$thisRegister}, $INITIALIZE_SETTINGS_METHOD_DESCRIPTOR
+                    move-result v$usableRegister
+                    if-eqz v$usableRegister, :notrevanced
                     return-void
-                """
+                """,
+                ExternalLabel("notrevanced", getInstruction(initializeSettingsIndex))
             )
         } ?: return AdPersonalizationActivityOnCreateFingerprint.toErrorResult()
 
         return PatchResultSuccess()
+    }
+
+    private fun String.toClassName(): String {
+        return substring(1, this.length - 1).replace("/", ".")
     }
 
     private companion object {
@@ -120,6 +105,11 @@ class TikTokSettingsPatch : BytecodePatch(
         private const val INITIALIZE_SETTINGS_METHOD_DESCRIPTOR =
             "$INTEGRATIONS_CLASS_DESCRIPTOR->initializeSettings(" +
                     "Lcom/bytedance/ies/ugc/aweme/commercialize/compliance/personalization/AdPersonalizationActivity;" +
-                    ")V"
+                    ")Z"
+        private const val CREATE_SETTINGS_ENTRY_METHOD_DESCRIPTOR =
+            "$INTEGRATIONS_CLASS_DESCRIPTOR->createSettingsEntry(" +
+                    "Ljava/lang/String;" +
+                    "Ljava/lang/String;" +
+                    ")Ljava/lang/Object;"
     }
 }