+++ src/main/kotlin/app/revanced/patches/youtube/layout/theme/resource/ThemeResourcePatch.kt
@@ -8,6 +8,7 @@ import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patches.shared.mapping.misc.patch.ResourceMappingPatch
 import app.revanced.patches.shared.settings.preference.impl.InputType
+import app.revanced.patches.shared.settings.preference.impl.StringResource
 import app.revanced.patches.shared.settings.preference.impl.TextPreference
 import app.revanced.patches.youtube.layout.seekbar.resource.SeekbarPreferencesPatch
 import app.revanced.patches.youtube.layout.theme.bytecode.patch.ThemeBytecodePatch.Companion.darkThemeBackgroundColor
@@ -19,10 +20,19 @@ import org.w3c.dom.Element
 class ThemeResourcePatch : ResourcePatch {
     override fun execute(context: ResourceContext): PatchResult {
         SeekbarPreferencesPatch.addPreferences(
+            SwitchPreference(
+                "revanced_seekbar_custom_color",
+                StringResource("revanced_seekbar_custom_color_title", "Enable custom seekbar color"),
+                StringResource("revanced_seekbar_custom_color_summary_on", "Custom seekbar color is shown"),
+                StringResource("revanced_seekbar_custom_color_summary_off", "Original seekbar color is shown")
+            ),
             TextPreference(
                 "revanced_seekbar_color",
                 "revanced_seekbar_color_title",
                 "revanced_seekbar_color_summary",
+//                "revanced_seekbar_custom_color_value",
+//                StringResource("revanced_seekbar_custom_color_value_title", "Custom seekbar color"),
+//                StringResource("revanced_seekbar_custom_color_value_summary", "The color of the seekbar"),
                 InputType.TEXT_CAP_CHARACTERS
             )
         )